@@ -2,7 +2,8 @@
 
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.22'
+gem 'rails', '4.0.13'
+gem 'protected_attributes'
 gem 'sqlite3'
 gem 'devise'
 gem 'dynamic_form'
@@ -11,15 +12,18 @@ gem 'actionmailer'
 
 gem 'rubocop'
 
-gem 'bundler-audit'
+# gem 'bundler-audit'
 
-group :cucumber do
-  gem 'database_cleaner'
-  gem 'cucumber-rails'
-  gem 'cucumber'
+group :development, :test do
+  gem 'pry-rails'
+  gem 'pry-nav'
 end
 
 group :test do
+  gem 'rspec-activemodel-mocks'
+  gem 'database_cleaner'
+  gem 'cucumber'
+  gem 'cucumber-rails', require: false
   gem 'test-unit' # delete later, this is a bug
   gem 'capybara'
   gem 'rspec', '2.99.0'
